Add ability to use data-hide-if as well as data-show-if for conditional statements
Why limit conditional statements to only trigger inputs to be shown when the option to hide them is just as useful at times. For example, I have an email address form send the visitor to a page with the full subscribe form where their email address is sent as a GET variable (auto-filling the field) which then doesn't need to be shown per them providing it in the previous step.

Very little code needed to be updated since showIf was extended to be showHideIf where you can pass which operation it should be doing. This also has the patched code from https://github.com/ibericode/mailchimp-for-wordpress/pull/375 implemented.