Replace gulp-cssmin with gulp-clean-css.
Due to vulnerabilities in the underlying clean-css library it needed to
be updated. This was challenging with gulp-cssmin as the project seems
to have been abandoned, so switch to the better maintained
gulp-clean-css for css minification.

The other change here is to update mithril, again to address a
vulnerability. It's updated to a major version, so this needs to be
tested closely to see if affected. Changelog can be seen here: https://github.com/MithrilJS/mithril.js/releases